+++ app/controllers/admins/tickets_controller.rb
@@ -57,10 +57,11 @@ class Admins::TicketsController < Admins::BaseController
   end
 
   def destroy_multiple
-    tickets = params[:tickets]
-    Ticket.where(id: tickets.keys).each do |ticket|
-      if !ticket.destroy
-        flash[:error] = ticket.errors.full_messages.join(". ")
+    if tickets = params[:tickets]
+      Ticket.where(id: tickets.keys).each do |ticket|
+        if !ticket.destroy
+          flash[:error] = ticket.errors.full_messages.join(". ")
+        end
       end
     end
     redirect_to admins_tickets_url
